silly workaround to make mypy happy
